feat: reserving ticket function done
