+++ src/hooks/api/useTicket.js
@@ -2,19 +2,19 @@ import useAsync from '../useAsync';
 import useToken from '../useToken';
 import * as ticketApi from '../../services/ticketApi';
 
-export default function useTicketType() {
+export default function useTicket() {
   const token = useToken();
   const {
-    data: ticketTypes,
-    loading: ticketTypesLoading,
-    error: ticketTypesError,
-    act: getTicketTypes,
-  } = useAsync(() => ticketApi.getTicketTypes(token));
+    data: ticket,
+    loading: ticketLoading,
+    error: ticketError,
+    act: createTicket,
+  } = useAsync((data) => ticketApi.createTicket(data, token));
 
   return {
-    ticketTypes,
-    ticketTypesLoading,
-    ticketTypesError,
-    getTicketTypes,
+    ticket,
+    ticketLoading,
+    ticketError,
+    createTicket,
   };
 }
