+++ src/hooks/api/useTicketType.js
@@ -1,0 +1,20 @@
+import useAsync from '../useAsync';
+import useToken from '../useToken';
+import * as ticketApi from '../../services/ticketApi';
+
+export default function useTicketType() {
+  const token = useToken();
+  const {
+    data: ticketTypes,
+    loading: ticketTypesLoading,
+    error: ticketTypesError,
+    act: getTicketTypes,
+  } = useAsync(() => ticketApi.getTicketTypes(token));
+
+  return {
+    ticketTypes,
+    ticketTypesLoading,
+    ticketTypesError,
+    getTicketTypes,
+  };
+}
